update directory structury, refactor manually
