+++ server/app/services/narrator.py
