+++ server/requirements.txt
@@ -1,13 +1,190 @@
+aiohappyeyeballs==2.6.1
+aiohttp==3.12.11
+aiosignal==1.3.2
+annotated-types==0.7.0
+anyio==4.9.0
+appdirs==1.4.4
+asgiref==3.8.1
+asttokens==3.0.0
+attrs==25.3.0
+auth0-python==4.9.0
+av==13.1.0
+backoff==2.2.1
+bcrypt==4.3.0
+beautifulsoup4==4.13.4
+blinker==1.9.0
+build==1.2.2.post1
+cachetools==5.5.2
+certifi==2025.4.26
+cffi==1.17.1
+charset-normalizer==3.4.2
+chroma-hnswlib==0.7.6
+chromadb==0.6.3
+click==8.2.1
+cloup==3.0.7
+colorama==0.4.6
+coloredlogs==15.0.1
 crewai==0.102.0
+cryptography==45.0.3
+dataclasses-json==0.6.7
+decorator==5.2.1
+distro==1.9.0
+docstring_parser==0.16
+durationpy==0.10
+et_xmlfile==2.0.0
+executing==2.2.0
 fastapi==0.115.11
-langchain_openai==0.3.7
+filelock==3.18.0
+flatbuffers==25.2.10
+frozenlist==1.6.2
+fsspec==2025.5.1
+glcontext==3.0.0
+google-auth==2.40.3
+googleapis-common-protos==1.70.0
+greenlet==3.2.3
+grpcio==1.72.1
+h11==0.16.0
+httpcore==1.0.9
+httptools==0.6.4
+httpx==0.27.2
+httpx-sse==0.4.0
+huggingface-hub==0.32.4
+humanfriendly==10.0
+idna==3.10
+importlib_metadata==8.7.0
+importlib_resources==6.5.2
+iniconfig==2.1.0
+instructor==1.8.3
+ipython==9.3.0
+ipython_pygments_lexers==1.1.1
+isosurfaces==0.1.2
+jedi==0.19.2
+Jinja2==3.1.6
+jiter==0.8.2
+json5==0.12.0
+json_repair==0.46.2
+jsonpatch==1.33
+jsonpickle==4.1.1
+jsonpointer==3.0.0
+jsonref==1.1.0
+jsonschema==4.24.0
+jsonschema-specifications==2025.4.1
+kubernetes==32.0.1
+langchain==0.3.25
+langchain-community==0.3.24
+langchain-core==0.3.64
+langchain-openai==0.3.7
+langchain-text-splitters==0.3.8
+langsmith==0.3.45
+litellm==1.60.2
+manim==0.19.0
+ManimPango==0.6.0
+mapbox_earcut==1.0.3
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+marshmallow==3.26.1
+matplotlib-inline==0.1.7
+mdurl==0.1.2
+mmh3==5.1.0
+moderngl==5.12.0
+moderngl-window==3.1.1
+mpmath==1.3.0
+multidict==6.4.4
+mutagen==1.47.0
+mypy_extensions==1.1.0
+networkx==3.5
+numpy==2.3.0
+oauthlib==3.2.2
+onnxruntime==1.22.0
+openai==1.84.0
+openpyxl==3.1.5
+opentelemetry-api==1.34.0
+opentelemetry-exporter-otlp-proto-common==1.34.0
+opentelemetry-exporter-otlp-proto-grpc==1.34.0
+opentelemetry-exporter-otlp-proto-http==1.34.0
+opentelemetry-instrumentation==0.55b0
+opentelemetry-instrumentation-asgi==0.55b0
+opentelemetry-instrumentation-fastapi==0.55b0
+opentelemetry-proto==1.34.0
+opentelemetry-sdk==1.34.0
+opentelemetry-semantic-conventions==0.55b0
+opentelemetry-util-http==0.55b0
+orjson==3.10.18
+overrides==7.7.0
+packaging==24.2
+parso==0.8.4
+pdfminer.six==20250327
+pdfplumber==0.11.6
+pillow==11.2.1
+pluggy==1.6.0
+posthog==4.4.0
+prompt_toolkit==3.0.51
+propcache==0.3.1
+protobuf==5.29.5
+pure_eval==0.2.3
+pyasn1==0.6.1
+pyasn1_modules==0.4.2
+pycairo==1.28.0
+pycparser==2.22
+pydantic==2.11.5
+pydantic-settings==2.9.1
+pydantic_core==2.33.2
+pydub==0.25.1
+pyglet==2.1.6
+pyglm==2.8.2
+Pygments==2.19.1
+PyJWT==2.10.1
+pypdf==5.6.0
+pypdfium2==4.30.1
+PyPika==0.48.9
+pyproject_hooks==1.2.0
+pyreadline3==3.5.4
+pytest==8.4.0
+python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
-Requests==2.32.3
+pyvis==0.3.2
+PyYAML==6.0.2
+referencing==0.36.2
+regex==2024.11.6
+requests==2.32.3
+requests-oauthlib==2.0.0
+requests-toolbelt==1.0.0
+rich==13.9.4
+rpds-py==0.25.1
+rsa==4.9.1
+scipy==1.15.3
+screeninfo==0.8.1
+shellingham==1.5.4
+six==1.17.0
+skia-pathops==0.8.0.post2
+sniffio==1.3.1
+soupsieve==2.7
+SQLAlchemy==2.0.41
+srt==3.5.3
+stack-data==0.6.3
+starlette==0.46.2
+svgelements==1.9.6
 sympy==1.13.3
+tenacity==9.1.2
+tiktoken==0.9.0
+tokenizers==0.21.1
+tomli==2.2.1
+tomli_w==1.2.0
+tqdm==4.67.1
+traitlets==5.14.3
+typer==0.16.0
+typing-inspect==0.9.0
+typing-inspection==0.4.1
+typing_extensions==4.14.0
+urllib3==2.4.0
+uv==0.7.12
 uvicorn==0.34.0
-pypdf
-langchain
-langchain-community
-chromadb
-pytest
-manim
+watchdog==6.0.0
+watchfiles==1.0.5
+wcwidth==0.2.13
+websocket-client==1.8.0
+websockets==15.0.1
+wrapt==1.17.2
+yarl==1.20.0
+zipp==3.22.0
+zstandard==0.23.0
